refactor(glyphs): drop the now-dead array_glyph _Unset sentinel
The apply_style refactor to DataStyle.for_apply_style removed the only
use of ArrayGlyph's private _Unset/_UNSET, leaving a second sentinel type
duplicating the canonical one in styling.params. Delete the dead
definition and relocate its repr/identity test onto params._Unset,
removing the latent cross-module footgun (passing one module's sentinel
into the other's isinstance gate).
